Added missing guard to get delegate

+++ Lynx/Schedules/ScheduleDetails/ScheduleDetailsVC.swift
@@ -40,6 +40,7 @@ class ScheduleDetailsVC: UIViewController {
     }
 
     @IBAction func saveSchedule(_ sender: Any) {
+        guard let delegate = detailDelegate else { return }
         if detailDelegate.isSaved() {
             self.didStartLoading()
             RestAPI.deleteSchedule(term: delegate.getTerm(),
